requirements.txt
pydantic>=2.0
transformers>=4.34.0
sentence-transformers>=2.2.2
chromadb>=0.4.15
google-genai>=1.0.0
torch>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
requests>=2.31.0
